feat: enable mini player expansion and unify audio services
